rh_service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==23.5
requests==2.31.0
gunicorn==21.2.0
djangorestframework-simplejwt
drf-yasg
django-extensions
uuid6

